ide code cleanup
