+++ src/main/java/systems/monomer/ide/EditorOutputStream.java
@@ -1,0 +1,37 @@
+package systems.monomer.ide;
+
+import org.jetbrains.annotations.Nullable;
+
+import javax.swing.*;
+import javax.swing.text.AttributeSet;
+import javax.swing.text.BadLocationException;
+import javax.swing.text.Document;
+import java.io.OutputStream;
+
+public class EditorOutputStream extends OutputStream {
+    private final StringBuilder stringBuilder = new StringBuilder();
+    private final JTextPane console;
+    private final @Nullable AttributeSet attributeSet;
+
+    EditorOutputStream(JTextPane console, @Nullable AttributeSet attributeSet) {
+        this.console = console;
+        this.attributeSet = attributeSet;
+    }
+
+    @Override
+    public void write(int b) {
+        stringBuilder.append((char) b);
+        if(b == '\n') {
+            String s = stringBuilder.toString();
+            stringBuilder.setLength(0);
+            SwingUtilities.invokeLater(() -> {
+                try {
+                    Document document = console.getDocument();
+                    document.insertString(document.getLength(), s, attributeSet);
+                } catch (BadLocationException e) {
+                    throw new RuntimeException(e);
+                }
+            });
+        }
+    }
+}
